requirements.txt
fastapi==0.128.0
uvicorn==0.32.0
motor==3.7.1
pymongo==4.9.2
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.12
passlib==1.7.4
bcrypt==5.0.0
email-validator==2.2.0
pydantic==2.12.5
pydantic-settings==2.13.0
dnspython==2.8.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
